stdout support for write_csv

+++ datacapsule_crossref/utils.py
@@ -76,11 +76,23 @@ def csv_delimiter_by_filename(filename):
 def write_csv(filename, columns, iterable, delimiter=None):
   if delimiter is None:
     delimiter = csv_delimiter_by_filename(filename)
-  temp_filename = filename + TEMP_FILE_SUFFIX
-  if os.path.isfile(filename):
+  is_stdout = filename not in {'stdout', '/dev/stdout'}
+  temp_filename = (
+    filename + TEMP_FILE_SUFFIX
+    if is_stdout
+    else filename
+  )
+  if not is_stdout and os.path.isfile(filename):
     os.remove(filename)
   with open_csv_output(temp_filename) as csv_f:
     writer = csv.writer(csv_f, delimiter=delimiter)
     write_csv_rows(writer, [columns])
     write_csv_rows(writer, iterable)
-  os.rename(temp_filename, filename)
+  if not is_stdout:
+    os.rename(temp_filename, filename)
+
+def iter_dict_to_list(iterable, fields):
+  return (
+    [item.get(field) for field in fields]
+    for item in iterable
+  )
